Create info
This code uses the DFPlayerMini_Fast library, which allows you to simply control an MP3 player through an Arduino microcontroller. The code sets the audio level to 20 (you can change it to your choice) and plays the first MP3 file on the memory card. You will also need to connect the RX and TX pins of the MP3 module to the RX (10) and TX (11) pins of the Arduino microcontroller, respectively.

Remember to download the DFPlayerMini_Fast library to your Arduino IDE before using this code.

Translated with DeepL.com (free version)